==> 03 Scripts/01 IC Data import and descriptive analysis.ipynb <==
# Content list 

1. Import librarie
2. Import data

# 01. Importlibraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

# to use a shortcut for importing data files, create a path (to project folder)
path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

# to create a list with the columns you need (excluding the ones you don't)
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# import orders dataset using the list previously created using the usecols argument. 
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# print first 5 rows of dataframe orders
df.head()

# import products dataset
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# print first 20 rows of dataframe products
df_prods.head(20)

# print last 35 rows of dataframe products
df_prods.tail(35)

# print name of columns
df_prods.columns

# print dimensions (number of rows and columns)
df_prods.shape

# print descriptive statistics
df_prods.describe()

# print data types
df_prods.dtypes

## the maximum value of aisle_id column is: 134
## the data type of the department_id column is: int64
