==> 03 Scripts/02 IC Data wrangling and subsetting.ipynb <==
# Content list

1. Importing libraries
2. Importing data
3. Data wrangling
4. Data dictionaries
5. Subsetting
6. Task 4.4
7. Exporting data

# 01.Importing libraries

import pandas as pd
import numpy as np
import os

# 02. Importing data

# Path to data files from project IC
path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

path

# Import dataset orders.csv  
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_ords.head()

df_ords.shape

# Import dataset products.csv  
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_prods.head()

df_prods.shape

# Import dataset departments.csv
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

# 03. Data wrangling

## 03.1 Dropping columns 

# Dropping eval_set column from orders.csv
df_ords.drop(columns = ['eval_set'])

# update the df_ords dataframe without the column eval_set
df_ords = df_ords.drop(columns = ['eval_set'])

df_ords.head()

# count values of specific columns (in this case 'days_since_prior_order column)
df_ords['days_since_prior_order'].value_counts(dropna = False)
# dropna = False is included to ensure that the NA values will also be counted and not excluded

## 03.2 Renaming columns

# Renaming column order_dow to orders_day_of_week
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

## 03.3 Changing variables data type

# Changing data type for column 'order_id' as strig for not to be consider when using the df.describe function 
df_ords['order_id'] = df_ords['order_id'].astype('str')

# To check if the change was made (df.dtype for single columns df.dtypes for multiple columns)
df_ords['order_id'].dtype
# 'O' is for object which means string

## 03.4 Transposing data

# Transposing df_dep (change rows into columns and vice versa)
df_dep.T

# Create a new version of the transposed version for df_dep
df_dep_t = df_dep.T

# check if the new vresion exists
df_dep_t
# the name of the columns are on the first row of dataframe not as headers

# to solve the issue of the headers
# first: add an index
df_dep_t.reset_index()

# second: create a new header by taking the first row of df_dep_t for the header
# iloc is for 'locate and integer'
new_header = df_dep_t.iloc[0]

new_header

# third: remove the first row to avoid having 2 rows 1 (one as a header and the other as row)
# it is done by creating a new dataframe that only copies the rows beyond the first row
df_dep_t_new = df_dep_t[1:]
# [1:] from row 1 (not row 0) and onward :

df_dep_t_new

# fourth: add the new_header to the new dataframe
df_dep_t_new.columns = new_header 

df_dep_t_new

# 04. Data dictionaries

# Create a data dictionary that contains meanings for the values in department_id column in the products dataframe
# the df_dep_t_new will be use as dictionary (because it contains the meaning for the department_id values)
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

# Use the data dictionary to determine what each number in the 'department_id' column stands for
print(data_dict.get('19'))

# 05. Subsetting

## Creating a smaller data set from a whole data set based on a particular filter.

# Creating a subsetf or df_prods dataframe that only contains data from the 'snacks' department. 
# indexing code that searches for the data in question 
df_prods['department_id'] == 19

# to return the values of products that correspond to the department_id number 19
df_prods[df_prods['department_id'] == 19]

# Save results into a new dataframe df_snacks
df_snacks = df_prods[df_prods['department_id'] == 19]

df_snacks.head()

# Another way to come to the same subset using loc (looking for column names)
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2.head()

# A third way using isin([19]) you are telling the loc function to look into a list
# It can be used when you need to use more than one value for example different department_ids (isin([17,18,19])
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3.head()

# 06. Task 4.4

## 06.2 Change data type of another variable from df_ords that doesn't need to be included as numeric variable

# change column order_number (it is the sequential number of orders) from df_ords
df_ords['order_number'] = df_ords['order_number'].astype('str')

df_ords['order_number'].dtype

## 06.3 Change the name of a variable in df_ords dataframe with unintuitive name

df_ords.head()

# I will change days_since_prior_order to days_since_last_order to use a more coloquial term, because I don't really think there's a 'unintuitive' name.
df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = True)

df_ords.head()

## 06.4 Find the busiest hour of the day for placing orders. 

# counting the values on the column named 'order_hour_of_day' to know which hour is the busiest one for placing orders.  
df_ords['order_hour_of_day'].value_counts(dropna=False)

The busiest hour for placing orders is 10 o'clock. 

## 06.5 Determine the meaning behind a value of 4 in the 'department_id' columns within the df_prods dataframe.

# using the data dictionary to find out what 4 means. 
print(data_dict.get('4'))

## 06.6 Create a subset containing information about breakfast items sales. 

# first step: find out which department_id correspond to 'breakfast'
data_dict

# department_id 14 is for breakfast
# second step: create a subset using a filter
df_breakfast = df_prods.loc[df_prods['department_id'] == 14]

df_breakfast.head()

## 06.7 Create subset including the following departments: alcohol, deli, beverages, and meat/seafood. 

# department ids: alcohol-5, deli-20, beverages-7, meat/seafood-12
# creating a subset using a filter for multiple columns)
df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinner_parties.head()

## 06.8 How many rows does the last dataframe you created have?

# dimensions of the dataframe
df_dinner_parties.shape

## 06.9 Information about user_id 1

# creating a subset of df_ords using the user_id 1 as filter
df_userid_1 = df_ords.loc[df_ords['user_id'] == 1]

df_userid_1.head()

## 06.10 Provide some detail about user_id 1 behavior.

# descriptive stats on subset df_userid_1
df_userid_1.describe()

The user place orders between Sunday and Wednesday.
Orders are placed between 7 and 16 hours. The 50% were placed between 7 and 8 in the morning. 
The average of days between orders is of 19 days, being the maximun 30 days and the minimum 0, which means that he had placed more than one order the same day. 


# 07. Exporting data

# Exporting dataframe df_orders
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'))

# Exporting dataframe df_dep_t_new
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
