==> 03 Scripts/03 IC Data consistency check.ipynb <==
# Content list

1. Importing librarires
2. Importing data
3. Data consistency check
4. Export data
5. Task 4.5
6. Exporting data

## 01. Importing libraries

# Importing pandas
import pandas as pd
import numpy as np
import os

## 02. Importing data

# Creating path to main folder
path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

# Import products.csv dataset
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_prods.head()

# Import orders_wrangled.csv dataset
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_ords.head()

# dropping unnamned column from orders dataframe
df_ords.drop(columns = ['Unnamed: 0'])

# updating df_ords
df_ords = df_ords.drop(columns = ['Unnamed: 0'])

df_ords.head()

## 03. Data Consistency Check

## 03.1 Mixed data types 

Example to deal with mixeddata types (Instacart has not mixed-type columns)

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True] 

df_test.head()

# function for checking whether a dataframe contains any mixed-type column
for col in df_test.columns.tolist(): # for loop (for this elements do this)
    weird = (df_test[[col]].applymap(type) != df_test [[col]].iloc[0].apply(type)).any(axis = 1) # variable weird with test that checks for 
#consistent datatypes within the columns. The weird variable take a boolean value of either True or False
# True (1) - column contains inconsistent data types
# False (0) - column contains only one data type
    if len (df_test[weird]) > 0: # if weird is true, then the command print is executed
        print(col)

df_test['mix'] = df_test['mix'].astype('str')

df_test.columns

## 03.2 Missing values

Finding missing values

# finding missing values in df_prods dataframe
df_prods.isnull().sum() # isnull function finds missing observations, then sum() the result of missing observations within each column

# create a subset of the dataframe containing only the null values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

Adressing missing values

Three ways with Python:
1. Create a new variable that acts like a flag based on the missing value.
2. Impute the value wiht the mean or median (only numeric values).
3. Remove or filter missing data.

# creating a new dataframe that excludes de null values (because string values can't be imputed like numeric values)
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

## False values

df_prods_clean.describe()

The max in prices is too big. I will check if there are other outlier prices. 

top_prices = df_prods.nlargest(20, 'prices')
top_prices

# Searching for other similar products to reduced fat milk
df_milk = df_prods_clean[df_prods_clean['product_name'].str.contains('Reduced Fat Milk')]
df_milk

There is another 2% reduced fat milk with a price of 9.2 so I will impute the false value with that. 

# Searching for reduced fat milk
df_cheese = df_prods_clean[df_prods_clean['product_name'].str.contains('Milkfat Cottage Cheese')]
df_cheese

I will use 7.7 to impute the price for lowfat 2% Milkfat Cottage Cheese

# Impute the new values for the false prices
df_prods_clean_imputed = df_prods_clean.copy()
df_prods_clean_imputed['prices'].replace({99999.0: 9.2, 14900.0: 7.7}, inplace=True)

df_prods_clean_imputed.describe()

## 03.3 Duplicate values

Finding duplicates

 # Creating a new subset 'df_dups' containing just duplicated values
df_dups = df_prods_clean_imputed[df_prods_clean_imputed.duplicated()]

df_dups

# creating new dataframe that excludes duplicates
df_prods_clean_no_dups = df_prods_clean_imputed.drop_duplicates()

df_prods_clean_no_dups.shape

df_prods_clean_no_dups.describe()

# 04. Exporting data

# Exporting dataframe df_prods_clean_no_dups 
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

# 05. Task 4.5

# Data consistency check on orders dataframe

df_ords.describe()

The column days_since_last_order has less counts than the other columns which means that there are missing values for this column or values with 0, which could also be the case for customers that placed more than one order on the same day.  
The days of the week goes from 0 to 6 which correspond to the days code 0 = saturday, 1 = sunday and so on. 
Orders are placed all day long between 00:00 and 23:00 hrs. The average hour for placing orders is 13:00. 
There's probably a problem on the order_number column, because the maximum number is 100, and there are a lot more orders placed.  

#looking at the inconsistency on the order_number column
df_ords.head(500)

There's definetly a problem with the order_number. For example order number 1, that is twice repeated do not correspond order_id or user_id.
** The order_number is repeated because it is a list of orders for each user_id. This means that the user with the most orders placed, had placed 100 orders. 

## 05.1 Mixed-type data

# looking for mixed-typed data in the orders dataframe
for col in df_ords.columns.tolist():
    weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_ords[weird]) > 0:
        print(col)

# looking for data types to double check
df_ords.info()

There are no mixed-type columns in the orders dataframe

## 05.2 Missing values

# looking for missing values on the orders dataframe
df_ords.isnull().sum()

df_ords.shape

# creating a column for the columns with missing values
df_ords_nan = df_ords[df_ords['days_since_last_order'].isnull() == True]

df_ords_nan

It seems like the first order for each user_id is NaN, which makes sense because there are not prior order. 

# looking for the count of unique user_id values to check if they correspond to the amount of rows with NaN data. 
df_ords['user_id'].nunique()

The number of rows with "missing data" correspond to the number of unique values for user_id, which means that the previous thought is true. The NaN values are in every first order for each customer_id. 

## Addressing missing values

To address the NaN values I will create a new column that will act as a flag based on the missing value. 

# creating a new dataframe
df_ords_clean = df_ords

# creating a new boolean column based on the missing values
df_ords_clean['new_customer'] = df_ords['days_since_last_order'].isnull() == True

df_ords_clean

## 05.3 Duplicate values

# looking for duplicates in the orders dataframe
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

There are no duplicate values. 

df_ords.shape

# 06. Exporting data

# Exporting dataframe df_ords 
df_ords_clean.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'))
