==> 03 Scripts/04.1 IC Combining and exporting data.ipynb <==
# Content list

1. Import libraries
2. Import data
3. Merge data
4. Export data

## 01. Import libraries

# Import libraries

import pandas as pd
import numpy as np
import os

## 02. Import data 

# Creating path to main folder

path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

# Import orders_products_prior dataset

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

df_ords_prior.head()

df_ords_prior.shape

# Import orders_checked dataset

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

df_ords.head()

df_ords.shape

## 03. Merge data

# Merge both dataframes

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

# Check if it's a full match

df_merged_large['_merge'].value_counts()

# Dropping 'Unnamned' column from df_merged_large dataframe

df_merged_large.drop(columns = ['Unnamed: 0'])

df_merged_large = df_merged_large.drop(columns = ['Unnamed: 0'])

df_merged_large.head()

df_merged_large.info()

df_merged_large.describe()

## 04. Export data

df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
