==> 03 Scripts/04.2 IC Combining and exporting data.ipynb <==
# Content list

1. Import libraries
2. Import data
3. Merge data
4. Export data

# 01. Import libraries

# Import libraries

import pandas as pd
import numpy as np
import os

# 02. Import data

# Creating path to main folder

path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

# Import orders_products_combined dataset

df_ords_prods_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_ords_prods_combined.shape

df_ords_prods_combined.head()

# Drop unnecessary column '_merge'

df_ords_prods_combined = df_ords_prods_combined.drop(columns = ['_merge'])

df_ords_prods_combined.head()

## Changing data type

df_ords_prods_combined.info()

# change data type'order_id'].astype('int32')
df_ords_prods_combined['order_id'] = df_ords_prods_combined['order_id'].astype('int32')
df_ords_prods_combined['user_id'] = df_ords_prods_combined['user_id'].astype('int32')
df_ords_prods_combined['order_number'] = df_ords_prods_combined['order_number'].astype('int8')
df_ords_prods_combined['orders_day_of_week'] = df_ords_prods_combined['orders_day_of_week'].astype('int8')
df_ords_prods_combined['order_hour_of_day'] = df_ords_prods_combined['order_hour_of_day'].astype('int8')
df_ords_prods_combined['days_since_last_order'] = df_ords_prods_combined['days_since_last_order'].astype('float32')
df_ords_prods_combined['product_id'] = df_ords_prods_combined['product_id'].astype('int32')
df_ords_prods_combined['add_to_cart_order'] = df_ords_prods_combined['add_to_cart_order'].astype('int32')
df_ords_prods_combined['reordered'] = df_ords_prods_combined['reordered'].astype('int8')

df_ords_prods_combined.info()

# Import products_checked dataset

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

df_prods.head()

# Drop unnecessary column 'Unnamed:0'

df_prods = df_prods.drop(columns = ['Unnamed: 0'])

df_prods.head()

df_prods.shape

df_prods.describe()

# 03. Merge data

I will use the merge function on 'product_id' column, with the default join (inner) to combine both datasets. 

# First test the merge

pd.merge(df_ords_prods_combined,df_prods, on = 'product_id', indicator = True)

# Merge both dataframes

df_merged = df_ords_prods_combined.merge(df_prods, on = 'product_id', indicator = True) 

df_merged.head()

df_merged.shape

# Count of '_merge' values

df_merged['_merge'].value_counts()

# drop '_merge' column
df_merged = df_merged.drop(columns = ['_merge'])

df_merged.head()

df_merged.shape

df_merged.info()

## Changing data type

df_merged.describe()

df_merged['aisle_id'] = df_merged['aisle_id'].astype('int32')
df_merged['department_id'] = df_merged['department_id'].astype('int8')
df_merged['prices'] = df_merged['prices'].astype('float32')

df_merged.info()

## 04. Export data

df_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
