==> 03 Scripts/05 IC Deriving new variables.ipynb <==
# Content list

1. Import libraries
2. Import data
3. Deriving new variables
4. Task 4.7
5. Export data

# 01. Import libraries

#Import libraries

import pandas as pd
import numpy as np
import os

# 02. Import data

# Create path to main folder

path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

# Import orders_products_merged dataset

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

ords_prods_merge.head()

# Create a subset with the first 1,000,000 rows

df = ords_prods_merge[:1000000]

df.shape

## 03. Deriving new variables

## 03.1 If statements with user-defined functions

# Define the function 'def' followed by the name 'price_label'  
# The argument (row) #looks at each row within the dataframe and : separates the head from the body

def price_label(row):  #

    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
          return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High-range product'
    else: return 'Not enough data'

# Apply the function

df['price_range'] = df.apply(price_label, axis=1)

# Count values for the new column

df['price_range'].value_counts(dropna = False)

df['prices'].max()

df.head()

## 03.2 If statements with the loc() function

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low range product'

df['price_range_loc'].value_counts(dropna = False)

df.head()

Creating the new column in the entire dataframe

ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

ords_prods_merge['price_range_loc'].value_counts(dropna = False)

## 03.3 If statements with for loops

# loop cycles for every value x within the range of 30 and 45
for x in range(30,45):
    print('My age is %d' % (x))

# Find out on which days most orders take place

ords_prods_merge['orders_day_of_week'].value_counts(dropna=False)

# Create the statement for a new column 'Busiest day'

result = [] # This is to create an empty list or shell for the results

for value in ords_prods_merge['orders_day_of_week']:
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else:
        result.append('Regularly busy')

result

# Create a new column to combine results with dataframe

ords_prods_merge['busiest_day'] = result

ords_prods_merge['busiest_day'].value_counts(dropna = False)

ords_prods_merge['busiest_day'].count()

ords_prods_merge.head()

# 04. Task 4.7

## 04.2 New column busiest days (plural)

# Create the statement for the new column

result_days = [] # This is to create an empty list or shell for the results

for value in ords_prods_merge['orders_day_of_week']:
    if value == 0 or value == 1:
        result_days.append('Busiest days')
    elif value == 3 or value == 4:
        result_days.append('Slowest days')
    else:
        result_days.append('Regular days')

result_days

# Create column for result_days

ords_prods_merge['busiest_days'] = result_days

# Frequency of new column

ords_prods_merge['busiest_days'].value_counts(dropna = False)

ords_prods_merge['busiest_days'].count()

ords_prods_merge.head()

## 04.3 Check the values for the new column

ords_prods_merge.count()

The column was succesfully created, the total counts of the column busiest_days match with the total counts of the previously created column busiest_day, and the total counts from the dataframe (32404859).  

## 04.4 Busiest hours of the day (periods of time)

# Find out which hours of the day have the most orders 

ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

Looking at the values, I decided to divide the periods of time according to the amount of orders: Most_orders will be the hours with >1,000,000 orders (10,11,14,15,13,12,16,9,17,8,18,19), Average_orders for the hours with >100,000 and <1,000,000 (20,7,21,22,23,6,0,1) and Fewest_orders will be the hours with >100,000 orders (5,2,4,3). 

# Create the statement for the new column

result_hours = []

# Define values group

group_most = [10,11,14,15,13,12,16,9,17,8,18,19]
group_average = [20,7,21,22,23,6,0,1]

# Define the loop function 

for value in ords_prods_merge['order_hour_of_day']:
    if value in group_most:
        result_hours.append('Most_orders')
    elif value in group_average:
        result_hours.append('Average_orders')
    else:
        result_hours.append('Fewest_orders')

result_hours

# Create column for result_days

ords_prods_merge['busiest_period_of_day'] = result_hours

# Frequency of new column

ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

ords_prods_merge.head()

# 05. Export data

ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_new_var.pkl'))
