==> 03 Scripts/07.2 IC Data visualizations.ipynb <==
# Content list
1. Import libraries
2. Import data
3. Descriptive Analysis
4. Wrangling data
5. Consistency check
6. Merge data
7. Export data

# 01. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

# Create path
path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

# import ords_prods_merge
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_new_var_customer_flags.pkl'))

ords_prods_merge.shape

# Import dataset customers
customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False) 

customers.head(30)

customers.shape

customers.tail(30)

customers.dtypes

# 03. Descriptive analysis

customers.describe()

The average age of customers is 50 years
Ages between 18 and 81
The customers have between 0 and 3 dependants
The min annual income is 26000 and the max is 600000

customers[['Gender', 'fam_status']].mode()

# 04. Wrangling data 

## 04.1 Rename columns 

# Renaming columns names to follow same format and logic
customers.rename(columns = {'First Name' : 'first_name', 'Surnam' : 'surname', 'Gender' : 'gender', 'STATE' : 'state', 'Age' : 'age', 'date_joined' : 'customer_since'}, inplace = True)

customers.columns

## 04.2 Drop columns

I will drop first_name and surname as they are not useful for the analysis and also for privacy issues could be necessary. 

# Create a new dataframe without the mentioned columns
customers_drop = customers.drop(['first_name', 'surname'], axis=1)

customers_drop.head()

## 04.3 Change datatype

I will change the datatype of customer_since to datetime and age and n_dependants to int8

# Change data type to datetime
customers_drop['customer_since'] = customers_drop['customer_since'].astype('datetime64[ns]')

# Change datatype to int8
customers_drop[['age', 'n_dependants']] = customers_drop[['age', 'n_dependants']].astype('int8')

customers_drop.dtypes

customers_drop.head()

Now I can include the dates in the descriptive analysis

customers_drop['customer_since'].describe()

The data is from clientes registered between 2017 and april 2020

# 05. Consistency check

## 05.1 Mixed data types

# Check for mixed daty-types columns
for col in customers_drop.columns.tolist():
    weird = (customers_drop[[col]].applymap(type) != customers_drop[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (customers_drop[weird]) > 0:
        print(col)

There are no mixed-typed columns

customers_drop.info()

## 05.2 Missing values

# Find missing values
customers_drop.isnull().sum()

There are no missing values

## 05.3 Duplicate values

# Create a new subset customers_dups' containing just duplicated values
customers_dups = customers_drop[customers_drop.duplicated()]

customers_dups

There are no duplicated values

customers_drop.shape

# 06. Merge data 

ords_prods_merge.dtypes

The key column is user_id, which in the customers dataset is type int64, and in the ords_prods_merge is ist32. Then I will change the dtype in the customers data. 

# Change data type for column user_id in customers dataframe
customers_drop['user_id'] = customers_drop['user_id'].astype('int32')

customers_drop.dtypes

# Merge both dataframes
ords_prods_all = ords_prods_merge.merge(customers_drop, on = 'user_id', indicator = True)

ords_prods_all.head()

# Check if it's a full match
ords_prods_all['_merge'].value_counts()

ords_prods_all.info()

# Remove '_merge' columnb
ords_prods_all = ords_prods_all.drop(['_merge'], axis=1)

ords_prods_all.head()

ords_prods_all.shape

# 07. Export data

# Export merged dataframe
ords_prods_all.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_all.pkl'))
