==> 03 Scripts/08.1 IC Data analysis and insights.ipynb <==
# Content list
1. Import libraries
2. Import data
3. Security implications
4. Analysis
5. Export data

# 01. Import libraries

# Import analysis and visualization libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 02. Import data

# Create path
path = r'C:\Users\adrig\OneDrive\Documentos\CARRIERFOUNDRY\DATA IMMERSION\ACHIEVEMENT 4\Ex_4.2_Python\03-2025 Instacart Basket Analysis'

# Import data
ords_prods_all = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_all.pkl'))

ords_prods_all.head(3)

# 03. Security implications 

PII data had been addressed in previous task (Notebook 4.9). 
Name and surname columns from customers were removed. 

# 04. Analysis

## 04.1 Comparison of customers behavior in different geographic areas. 

# Check column state
ords_prods_all['state'].nunique()

ords_prods_all['state'].value_counts(dropna=False)

# Define the regions
northeast = ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey']
midwest = ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']
south = ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana']
west = ['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']

# Assign region values using loc() and isin()
ords_prods_all.loc[ords_prods_all['state'].isin(northeast), 'region'] = 'Northeast'

ords_prods_all.loc[ords_prods_all['state'].isin(midwest), 'region'] = 'Midwest'

ords_prods_all.loc[ords_prods_all['state'].isin(south), 'region'] = 'South'

ords_prods_all.loc[ords_prods_all['state'].isin(west), 'region'] = 'West'

# Check new column
ords_prods_all['region'].value_counts()

# Crosstab to compare spending habits by region
crosstab_spending_region = pd.crosstab(ords_prods_all['spending_flag'], ords_prods_all['region'], dropna=False)
crosstab_spending_region

All regions have a very small proportion of high spenders. 
The South region has the highest count of both high and low spender, suggesting that it has the larger customer base. 

# Crosstab to compare spending habits by loyalty status
crosstab_spending_loyal = pd.crosstab(ords_prods_all['spending_flag'], ords_prods_all['loyalty_flag'], dropna=False)
crosstab_spending_loyal

## 04.2 Exclusion flag

# Create an exclusion flag for low-activity customers
ords_prods_all.loc[ords_prods_all['max_order'] < 5, 'exclusion_flag'] = 'Yes'

ords_prods_all.loc[ords_prods_all['max_order'] >= 5, 'exclusion_flag'] = 'No'

# Check frequency of values 
ords_prods_all['exclusion_flag'].value_counts(dropna=False)

ords_prods_all[['user_id', 'max_order', 'exclusion_flag']].sort_values('max_order').head(30)

ords_prods_all.groupby('exclusion_flag')['user_id'].nunique()

43,578 customers are low-activity customers

# Remove observations from low_activity customers by creating a new dataframe
ords_prods_all_ex = ords_prods_all.drop(ords_prods_all[ords_prods_all['exclusion_flag'] =='Yes'].index)

ords_prods_all_ex.shape

ords_prods_all.shape

len(ords_prods_all) - len(ords_prods_all_ex)

# Remove exclusion flag
ords_prods_all_ex = ords_prods_all_ex.drop(columns = ['exclusion_flag'])

ords_prods_all_ex.shape

# 05 Export data

# Export new dataframe with exclusions
ords_prods_all_ex.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_all_exclusions.pkl'))
